requirements.txt
# Core
requests>=2.31.0
python-dotenv>=1.0.0

# UI
PySide6>=6.6.0

# Database
SQLAlchemy>=2.0.0

# Scheduler
APScheduler>=3.10.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0
responses>=0.23.0
ruff>=0.1.0
mypy>=1.5.0
